--- a/src/system.h
+++ b/src/system.h
@@ -5,6 +5,7 @@
 
 #include "alloc.h"
 #include "execution.h"
+#include "monopole.h"
 #include "vec.h"
 
 template <typename T, dim_t N>
@@ -15,8 +16,8 @@ class System {
   index_t const max_tree_node_size;
   T const dt;
   T const constant;
-  vector<T> m;
-  vector<vec<T, N>> x, v, a, ao;
+  vector<monopole<T, N>> p;
+  vector<vec<T, N>> v, a, ao;
 
   // random generation
   std::mt19937 gen{42};  // fix random generation
@@ -29,8 +30,7 @@ class System {
      max_tree_node_size(std::max<std::size_t>(child_count<N> * size, 1000)),
      dt(time_step),
      constant(constant),
-     m(size),
-     x(size),
+     p(size),
      v(size),
      a(size),
      ao(size) {}
@@ -39,21 +39,21 @@ class System {
 
   // Helper to make it easier to access all state from parallel algorithms
   struct state_t {
-    T* m;
-    vec<T, N>*x, *v, *a, *ao;
+    monopole<T, N>* p;
+    vec<T, N>*v, *a, *ao;
     T dt, c;
     index_t sz;
   };
   state_t state() {
     return {
-     .m = m.data(), .x = x.data(), .v = v.data(), .a = a.data(), .ao = ao.data(), .dt = dt, .c = constant, .sz = size};
+     .p = p.data(), .v = v.data(), .a = a.data(), .ao = ao.data(), .dt = dt, .c = constant, .sz = size};
   }
 
   void accelerate_step() {
     // performs leap frog integration
     auto r = body_indices();
     std::for_each(par_unseq, r.begin(), r.end(), [s = state()](auto i) {
-      s.x[i] += s.dt * s.v[i] + T(0.5) * s.dt * s.dt * s.ao[i];
+      s.p[i].x() += s.dt * s.v[i] + T(0.5) * s.dt * s.dt * s.ao[i];
       s.v[i] += T(0.5) * s.dt * (s.a[i] + s.ao[i]);
       s.ao[i] = s.a[i];
     });
@@ -63,15 +63,16 @@ class System {
     auto r           = body_indices();
     T kinetic_energy = static_cast<T>(0.5)
                        * std::transform_reduce(par_unseq, r.begin(), r.end(), static_cast<T>(0), std::plus<T>{},
-                                               [m = m.data(), v = v.data()](auto i) { return m[i] * l2norm2(v[i]); });
+                                               [p = p.data(), v = v.data()](auto i) { return p[i].mass() * l2norm2(v[i]); });
     T gravitational_energy = -static_cast<T>(0.5) * constant
                              * std::transform_reduce(par_unseq, r.begin(), r.end(), static_cast<T>(0), std::plus<T>{},
-                                                     [m = m.data(), x = x.data(), sz = size](auto i) {
+                                                     [p = p.data(), sz = size](auto i) {
                                                        T total = 0;
-                                                       T mi    = m[i];
-                                                       auto xi = x[i];
-                                                       for (index_t j = 0; j < sz; j++)
-                                                         if (j != i) total += mi * m[j] / dist(xi, x[j]);
+                                                       auto [mi, xi] = p[i];
+                                                       for (index_t j = 0; j < sz; j++) {
+							 auto [mj, xj] = p[j];
+                                                         if (j != i) total += mi * mj / dist(xi, xj);
+						       }
                                                        return total;
                                                      });
 
@@ -82,8 +83,7 @@ class System {
     auto i = next_point;
     next_point += 1;
 
-    m[i] = mass;
-    x[i] = pos;
+    p[i] = monopole(mass, pos);
     v[i] = vel;
   }
 
@@ -91,7 +91,7 @@ class System {
     for (size_t i = 0; i < size; i++) {
       std::cout << std::format("{:02}: m={: .3e}, p=({: .3e}, {: .3e}), v=({: "
                                ".3e}, {: .3e}), f=({: .3e}, {: .3e})",
-                               i, m[i], x[i][0], x[i][1], v[i][0], v[i][1], a[i][0], a[i][1])
+                               i, p[i].mass(), p[i].x()[0], p[i].x()[1], v[i][0], v[i][1], a[i][0], a[i][1])
                 << std::endl;
     }
   }