@illuhad 2024-08-14
Tested it on AMD and NVIDIA:

# RTX A5000:

## Before

all-pairs (70k particles): 69.5ms/step
all-pairs-collapsed (70k particles): 24.9ms/step
octree(7M particles): 1735ms/step
bvh(7M particles): 1296ms/step

## After

all-pairs (70k particles): 67.9ms/step
all-pairs-collapsed (70k particles): 25.1ms/step
octree(7M particles): 1696ms/step
bvh(7M particles): 1279ms/step


# AMD Radeon Pro VII:

## Before

all-pairs (70k particles): 221.3ms/step
bvh(7M particles): 6391ms/step

## After

all-pairs (70k particles): 221.8 ms/step
bvh(7M particles): 6295ms/step



It seems to work, but the perf benefit is only very small (1-2%), even for all-pairs.