@gonzalobg 2024-08-14
Thanks, similar results on my end, negligible impact. 
Should we merge this?