@illuhad 2024-08-14
I'm not sure. Do you feel like it simplifies or complicates code?

One thing I'm having a hard time understanding is why packing made such a big difference for the monopoles, but not for the particles, even in all-pairs which only relies on loading particles. Is it possible that perhaps something is not working right?